feat: add `getEffectiveDepositCap`

+++ packages/core-contracts/test/StakingRewardsManagerBase.t.sol
@@ -2,14 +2,16 @@
 pragma solidity 0.8.28;
 
 import {FleetStakingRewardsManager} from "../src/contracts/FleetStakingRewardsManager.sol";
-import {IFleetStakingRewardsManager} from "../src/interfaces/IFleetStakingRewardsManager.sol";
+
+import {ProtocolAccessManager} from "../src/contracts/ProtocolAccessManager.sol";
 import {IStakingRewardsManagerBaseErrors} from "../src/errors/IStakingRewardsManagerBaseErrors.sol";
+import {IFleetStakingRewardsManager} from "../src/interfaces/IFleetStakingRewardsManager.sol";
+
+import {IProtocolAccessManager} from "../src/interfaces/IProtocolAccessManager.sol";
 import {MockSummerGovernor} from "./mocks/MockSummerGovernor.sol";
+import {ERC20Mock} from "@openzeppelin/contracts/mocks/token/ERC20Mock.sol";
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {Test, console} from "forge-std/Test.sol";
-import {ERC20Mock} from "@openzeppelin/contracts/mocks/token/ERC20Mock.sol";
-import {ProtocolAccessManager} from "../src/contracts/ProtocolAccessManager.sol";
-import {IProtocolAccessManager} from "../src/interfaces/IProtocolAccessManager.sol";
 
 contract StakingRewardsManagerBaseTest is Test {
     FleetStakingRewardsManager public stakingRewardsManager;
@@ -32,7 +34,7 @@ contract StakingRewardsManagerBaseTest is Test {
 
         // Deploy mock tokens
         console.log("Deploying mock tokens");
-        for (uint i = 0; i < 3; i++) {
+        for (uint256 i = 0; i < 3; i++) {
             rewardTokens.push(new ERC20Mock());
         }
 
@@ -45,7 +47,7 @@ contract StakingRewardsManagerBaseTest is Test {
         address[] memory rewardTokenAddresses = new address[](
             rewardTokens.length
         );
-        for (uint i = 0; i < rewardTokens.length; i++) {
+        for (uint256 i = 0; i < rewardTokens.length; i++) {
             rewardTokenAddresses[i] = address(rewardTokens[i]);
         }
         IProtocolAccessManager accessManager = new ProtocolAccessManager(
@@ -60,7 +62,7 @@ contract StakingRewardsManagerBaseTest is Test {
         console.log("Minting initial tokens");
         mockFleetCommander.mint(alice, INITIAL_STAKE_AMOUNT);
         mockFleetCommander.mint(bob, INITIAL_STAKE_AMOUNT);
-        for (uint i = 0; i < rewardTokens.length; i++) {
+        for (uint256 i = 0; i < rewardTokens.length; i++) {
             rewardTokens[i].mint(
                 address(stakingRewardsManager),
                 INITIAL_REWARD_AMOUNT
@@ -360,7 +362,7 @@ contract StakingRewardsManagerBaseTest is Test {
 
         // Notify rewards for all three tokens
         vm.startPrank(address(mockGovernor));
-        for (uint i = 0; i < rewardTokens.length; i++) {
+        for (uint256 i = 0; i < rewardTokens.length; i++) {
             stakingRewardsManager.notifyRewardAmount(
                 IERC20(address(rewardTokens[i])),
                 rewardAmounts[i],
@@ -374,7 +376,7 @@ contract StakingRewardsManagerBaseTest is Test {
 
         // Check earned amounts
         uint256[] memory earnedAmounts = new uint256[](3);
-        for (uint i = 0; i < rewardTokens.length; i++) {
+        for (uint256 i = 0; i < rewardTokens.length; i++) {
             earnedAmounts[i] = stakingRewardsManager.earned(
                 alice,
                 IERC20(address(rewardTokens[i]))
@@ -405,14 +407,14 @@ contract StakingRewardsManagerBaseTest is Test {
         uint256[] memory balancesBefore = new uint256[](3);
         uint256[] memory balancesAfter = new uint256[](3);
 
-        for (uint i = 0; i < rewardTokens.length; i++) {
+        for (uint256 i = 0; i < rewardTokens.length; i++) {
             balancesBefore[i] = rewardTokens[i].balanceOf(alice);
         }
 
         vm.prank(alice);
         stakingRewardsManager.getReward();
 
-        for (uint i = 0; i < rewardTokens.length; i++) {
+        for (uint256 i = 0; i < rewardTokens.length; i++) {
             balancesAfter[i] = rewardTokens[i].balanceOf(alice);
             assertGt(
                 balancesAfter[i],
@@ -440,7 +442,7 @@ contract StakingRewardsManagerBaseTest is Test {
         vm.warp(block.timestamp + 3.5 days);
 
         // Check final earned amounts
-        for (uint i = 0; i < rewardTokens.length; i++) {
+        for (uint256 i = 0; i < rewardTokens.length; i++) {
             uint256 finalEarnedAmount = stakingRewardsManager.earned(
                 alice,
                 IERC20(address(rewardTokens[i]))
@@ -471,7 +473,7 @@ contract StakingRewardsManagerBaseTest is Test {
         uint256[] memory balancesBeforeFinal = new uint256[](
             rewardTokens.length
         );
-        for (uint i = 0; i < rewardTokens.length; i++) {
+        for (uint256 i = 0; i < rewardTokens.length; i++) {
             balancesBeforeFinal[i] = rewardTokens[i].balanceOf(alice);
         }
 
@@ -480,7 +482,7 @@ contract StakingRewardsManagerBaseTest is Test {
         stakingRewardsManager.getReward();
 
         // Verify all rewards have been claimed
-        for (uint i = 0; i < rewardTokens.length; i++) {
+        for (uint256 i = 0; i < rewardTokens.length; i++) {
             uint256 finalBalance = rewardTokens[i].balanceOf(alice);
             uint256 rewardReceived = finalBalance - balancesBeforeFinal[i];
